Revert to previous unmount behavior
In order to unmount a volume after the plugin is restarted, we want to
attempt a limited unmount when api calls for it.  This change is to
revert to that behavior instead of ignoring the unmount.  This prevents
needing to restart docker and every container using rbd-docker-plugin
during plugin upgrades.

Now the newly started rbd-docker-plugin will still attempt to unmount
when requested, even if the volume is not in memory.  This will still
only work from the same host (needed when original container using
volume stops).  Any other use case may need manual intervention